refactor: добавит слайсы
