+++ project/src/store/middlewares/redirect.ts
@@ -1,8 +1,8 @@
 import { browserHistory } from '../../browser-history';
 import { Middleware } from 'redux';
-import { reducer } from '../reducer';
+import { rootReducer } from '../root-reducer';
 
-type Reducer = ReturnType<typeof reducer>;
+type Reducer = ReturnType<typeof rootReducer>;
 
 export const redirect: Middleware<unknown, Reducer> = (_store) => (next) => (action) => {
   if (action.type === 'main/redirectToRoute') {
